move import blocks to blockchain struct
